+++ patches/Accessibility/strings_de.cpp
@@ -142,6 +142,7 @@ const char* Get(Id id) {
         case Id::FmtEquipSlotItem:     return "%s, %s";
         case Id::FmtEquipSlotEmpty:    return "%s, leer";
         case Id::EquipUnequipped:      return "Ausr\xFCstung abgelegt";
+        case Id::FmtEquipStillEquipped: return "Gegenstand weiterhin angelegt: %s";
         case Id::FmtWeaponSwitchedOne: return "Waffen gewechselt: %s";
         case Id::FmtWeaponSwitchedTwo: return "Waffen gewechselt: %s und %s";
         case Id::WeaponSwitchedBare:   return "Waffen gewechselt";