+++ patches/Accessibility/strings_fr.cpp
@@ -139,6 +139,7 @@ const char* Get(Id id) {
         case Id::FmtEquipSlotItem:     return "%s, %s";
         case Id::FmtEquipSlotEmpty:    return "%s, vide";
         case Id::EquipUnequipped:      return "\xC9quipement retir\xE9";
+        case Id::FmtEquipStillEquipped: return "Objet toujours \xE9quip\xE9 : %s";
         case Id::FmtWeaponSwitchedOne: return "Armes chang\xE9""es : %s";
         case Id::FmtWeaponSwitchedTwo: return "Armes chang\xE9""es : %s et %s";
         case Id::WeaponSwitchedBare:   return "Armes chang\xE9""es";